New version
I have reduced the script in some ways and increased it in others. Added Detection for resolution, if the file is 1920x1080 it doesn't matter, if it's 960x540 and 16:9 then it'll upscale it to 1024x576 to interlace properly. If you're video is 4:3 within 16:9 frame then add the argument -4:3 and it'll crop it and make it 768x576.